requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
qrcode==8.2
pillow>=11.0.0
python-dotenv==1.0.0
alembic==1.13.0
python-jose==3.3.0
passlib>=1.7
argon2-cffi>=23.1.0
python-multipart>=0.0.6
